Create gameBoard module
The project prompt says to use an array for the game board inside of
the module, but it seems like it would be easier to check for the win
condition if an object is used instead. There's only 9 spaces and 8
possible win conditions, so I'm not dealing with a lot of data here.