Create _checkForWin
This function feels like a cluster, but it's the best I could come up
with after thinking about it and searching online for a while.

This also means the board will need to be a two-dimensional array, as
originally hinted at by the prompt, and my other functions will need to
be changed.