Add enableBoard and _placeToken, refactor _board
Of course it should be no surprise that I was overthinking the simplest
part of the script by turning the array into an object, then a 2d array
and then eventually to just back to an array... Occam's Razor I guess.

But this wouldn't be an educational experience if I wasn't screwing up.