Refactor _board and _placeToken. Not working ATM.
I feel like Jimmy Neutron had a brain blast with this idea but then
feel like Charlie after uncovering a company conspiracy when I think
about having explaining it.

I can just make _board an array of objects for each slot which contains
all of the relevant properties needed for checking which slot has been
clicked on, updating the relevant element in the DOM, tracking what
token is in each slot and what rows/columns/diagonals need to be
checked for a win condition.

We'll see how well this turns out.